test(cli): split CRUD business-logic suites into focused modules
