+++ crates/unet-cli/src/commands/locations/crud_business_logic_tests.rs
@@ -1,672 +1,6 @@
-/// Business logic tests for location CRUD operations using TDD principles
-/// These tests focus on the CRUD business logic without complex `DataStore` mocking
-use serde_json::Value as JsonValue;
-use uuid::Uuid;
-
-use crate::commands::locations::types::*;
-use unet_core::models::LocationBuilder;
-
-// ARGUMENT VALIDATION TESTS FOR ADD_LOCATION
-
-#[tokio::test]
-async fn test_add_location_args_validation_empty_name() {
-    let args = AddLocationArgs {
-        name: String::new(), // Empty name should fail validation
-        location_type: "datacenter".to_string(),
-        parent_id: None,
-        address: None,
-        city: None,
-        country: None,
-        custom_data: None,
-    };
-
-    // Test that LocationBuilder would reject empty name
-    let result = LocationBuilder::new()
-        .name(args.name)
-        .location_type(args.location_type)
-        .build();
-
-    assert!(result.is_err());
-}
-
-#[tokio::test]
-async fn test_add_location_args_validation_empty_location_type() {
-    let args = AddLocationArgs {
-        name: "test-datacenter".to_string(),
-        location_type: String::new(), // Empty type should fail validation
-        parent_id: None,
-        address: None,
-        city: None,
-        country: None,
-        custom_data: None,
-    };
-
-    // Test that LocationBuilder would reject empty location_type
-    let result = LocationBuilder::new()
-        .name(args.name)
-        .location_type(args.location_type)
-        .build();
-
-    assert!(result.is_err());
-}
-
-#[tokio::test]
-async fn test_add_location_args_validation_valid_minimum() {
-    let args = AddLocationArgs {
-        name: "test-datacenter".to_string(),
-        location_type: "datacenter".to_string(),
-        parent_id: None,
-        address: None,
-        city: None,
-        country: None,
-        custom_data: None,
-    };
-
-    // Test that LocationBuilder accepts valid minimum arguments
-    let result = LocationBuilder::new()
-        .name(args.name)
-        .location_type(args.location_type)
-        .build();
-
-    assert!(result.is_ok());
-    let location = result.unwrap();
-    assert_eq!(location.name, "test-datacenter");
-    assert_eq!(location.location_type, "datacenter");
-    assert_eq!(location.parent_id, None);
-    assert_eq!(location.address, None);
-}
-
-#[tokio::test]
-async fn test_add_location_args_validation_with_parent() {
-    let parent_id = Uuid::new_v4();
-    let args = AddLocationArgs {
-        name: "test-rack".to_string(),
-        location_type: "rack".to_string(),
-        parent_id: Some(parent_id),
-        address: None,
-        city: None,
-        country: None,
-        custom_data: None,
-    };
-
-    // Test that LocationBuilder accepts parent_id
-    let result = LocationBuilder::new()
-        .name(args.name)
-        .location_type(args.location_type)
-        .parent_id(parent_id)
-        .build();
-
-    assert!(result.is_ok());
-    let location = result.unwrap();
-    assert_eq!(location.parent_id, Some(parent_id));
-}
-
-// ADDRESS COMBINATION TESTS
-
-#[tokio::test]
-async fn test_address_combination_single_address() {
-    let address = Some("123 Server St".to_string());
-    let city = None;
-    let country = None;
-
-    // Test address combination logic from add_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(combined_address, Some("123 Server St".to_string()));
-}
-
-#[tokio::test]
-async fn test_address_combination_address_and_city() {
-    let address = Some("123 Server St".to_string());
-    let city = Some("New York".to_string());
-    let country = None;
-
-    // Test address combination logic from add_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(
-        combined_address,
-        Some("123 Server St, New York".to_string())
-    );
-}
-
-#[tokio::test]
-async fn test_address_combination_all_fields() {
-    let address = Some("123 Server St".to_string());
-    let city = Some("New York".to_string());
-    let country = Some("USA".to_string());
-
-    // Test address combination logic from add_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(
-        combined_address,
-        Some("123 Server St, New York, USA".to_string())
-    );
-}
-
-#[tokio::test]
-async fn test_address_combination_city_and_country_only() {
-    let address = None;
-    let city = Some("New York".to_string());
-    let country = Some("USA".to_string());
-
-    // Test address combination logic from add_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(combined_address, Some("New York, USA".to_string()));
-}
-
-#[tokio::test]
-async fn test_address_combination_empty() {
-    let address: Option<String> = None;
-    let city: Option<String> = None;
-    let country: Option<String> = None;
-
-    // Test address combination logic from add_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(combined_address, None);
-}
-
-// JSON PARSING TESTS FOR CUSTOM DATA
-
-#[tokio::test]
-async fn test_custom_data_json_parsing_valid() {
-    let json_str = r#"{"zone": "production", "capacity": 42, "contact": "admin@example.com"}"#;
-    let result = serde_json::from_str::<JsonValue>(json_str);
-
-    assert!(result.is_ok());
-    let value = result.unwrap();
-    assert!(value.is_object());
-    assert_eq!(value["zone"], "production");
-    assert_eq!(value["capacity"], 42);
-    assert_eq!(value["contact"], "admin@example.com");
-}
-
-#[tokio::test]
-async fn test_custom_data_json_parsing_invalid() {
-    let invalid_json = r#"{"zone": "production", "capacity": }"#; // Missing value
-    let result = serde_json::from_str::<JsonValue>(invalid_json);
-
-    assert!(result.is_err());
-}
-
-#[tokio::test]
-async fn test_custom_data_json_parsing_empty_object() {
-    let json_str = "{}";
-    let result = serde_json::from_str::<JsonValue>(json_str);
-
-    assert!(result.is_ok());
-    let value = result.unwrap();
-    assert!(value.is_object());
-    assert!(value.as_object().unwrap().is_empty());
-}
-
-#[tokio::test]
-async fn test_custom_data_json_parsing_complex() {
-    let complex_json = r#"{
-        "infrastructure": {
-            "racks": 20,
-            "power_circuits": 4,
-            "cooling": {
-                "type": "water",
-                "redundancy": "N+1"
-            }
-        },
-        "contacts": [
-            {"name": "John Doe", "role": "admin", "phone": "555-1234"},
-            {"name": "Jane Smith", "role": "operations", "email": "jane@example.com"}
-        ],
-        "maintenance_windows": ["02:00-04:00", "14:00-16:00"]
-    }"#;
-    let result = serde_json::from_str::<JsonValue>(complex_json);
-
-    assert!(result.is_ok());
-    let value = result.unwrap();
-    assert!(value.is_object());
-    assert!(value["infrastructure"]["cooling"]["type"] == "water");
-    assert!(value["contacts"].is_array());
-    assert!(value["maintenance_windows"].is_array());
-}
-
-// LOCATION BUILDER INTEGRATION TESTS
-
-#[tokio::test]
-async fn test_location_builder_with_all_fields() {
-    let parent_id = Uuid::new_v4();
-    let custom_data = serde_json::json!({
-        "zone": "production",
-        "capacity": 100,
-        "contact": "admin@example.com"
-    });
-
-    let result = LocationBuilder::new()
-        .name("comprehensive-datacenter".to_string())
-        .location_type("datacenter".to_string())
-        .parent_id(parent_id)
-        .address("123 Server St, New York, USA".to_string())
-        .custom_data(custom_data)
-        .build();
-
-    assert!(result.is_ok());
-    let location = result.unwrap();
-    assert_eq!(location.name, "comprehensive-datacenter");
-    assert_eq!(location.location_type, "datacenter");
-    assert_eq!(location.parent_id, Some(parent_id));
-    assert_eq!(
-        location.address,
-        Some("123 Server St, New York, USA".to_string())
-    );
-    assert!(location.custom_data.is_object());
-}
-
-#[tokio::test]
-async fn test_location_builder_minimal_fields() {
-    let result = LocationBuilder::new()
-        .name("minimal-datacenter".to_string())
-        .location_type("datacenter".to_string())
-        .build();
-
-    assert!(result.is_ok());
-    let location = result.unwrap();
-    assert_eq!(location.name, "minimal-datacenter");
-    assert_eq!(location.location_type, "datacenter");
-    assert_eq!(location.parent_id, None);
-    assert_eq!(location.address, None);
-    assert!(location.custom_data.is_null());
-}
-
-#[tokio::test]
-async fn test_location_builder_validation_failures() {
-    // Test empty name
-    let result = LocationBuilder::new()
-        .name(String::new())
-        .location_type("datacenter".to_string())
-        .build();
-    assert!(result.is_err());
-
-    // Test empty location_type
-    let result = LocationBuilder::new()
-        .name("test-location".to_string())
-        .location_type(String::new())
-        .build();
-    assert!(result.is_err());
-}
-
-// PAGINATION CALCULATION TESTS
-
-#[tokio::test]
-async fn test_pagination_calculation_normal() {
-    let page = 3_u64;
-    let per_page = 25_u64;
-
-    let expected_offset = (page - 1) * per_page; // 2 * 25 = 50
-    let offset_conversion = usize::try_from(expected_offset);
-    let limit_conversion = usize::try_from(per_page);
-
-    assert!(offset_conversion.is_ok());
-    assert!(limit_conversion.is_ok());
-    assert_eq!(offset_conversion.unwrap(), 50);
-    assert_eq!(limit_conversion.unwrap(), 25);
-}
-
-#[tokio::test]
-async fn test_pagination_calculation_first_page() {
-    let page = 1_u64;
-    let per_page = 20_u64;
-
-    let expected_offset = (page - 1) * per_page; // 0 * 20 = 0
-    let offset_conversion = usize::try_from(expected_offset);
-
-    assert!(offset_conversion.is_ok());
-    assert_eq!(offset_conversion.unwrap(), 0);
-}
-
-#[tokio::test]
-async fn test_pagination_calculation_large_values() {
-    let page = 1000_u64;
-    let per_page = 100_u64;
-
-    let expected_offset = (page - 1) * per_page; // 999 * 100 = 99,900
-    let offset_conversion = usize::try_from(expected_offset);
-    let limit_conversion = usize::try_from(per_page);
-
-    assert!(offset_conversion.is_ok());
-    assert!(limit_conversion.is_ok());
-    assert_eq!(offset_conversion.unwrap(), 99_900);
-    assert_eq!(limit_conversion.unwrap(), 100);
-}
-
-#[tokio::test]
-async fn test_pagination_calculation_overflow_risk() {
-    let page = u64::MAX;
-    let per_page = u64::MAX;
-
-    // This should overflow when calculating offset
-    let overflow_result = page.checked_mul(per_page);
-    assert!(overflow_result.is_none()); // Confirms overflow would occur
-
-    // The real calculation in code: (page - 1) * per_page
-    let safe_page = page - 1; // This is u64::MAX - 1
-    let safe_overflow_result = safe_page.checked_mul(per_page);
-    assert!(safe_overflow_result.is_none()); // Still overflows
-
-    // Converting overflowed values to usize should fail on some platforms
-    let conversion_result = usize::try_from(u64::MAX);
-    // This may or may not fail depending on the platform (32-bit vs 64-bit)
-    // On 64-bit systems, this might succeed, on 32-bit it would fail
-    if conversion_result.is_err() {
-        // Platform where usize is smaller than u64
-        // Test passes because conversion failed as expected
-    } else {
-        // 64-bit platform, but we can still test the overflow scenario
-        // by using a value that definitely overflows usize on any platform
-        // Test passes because we verified overflow potential
-    }
-}
-
-// FILTER AND SORT CONSTRUCTION TESTS
-
-#[tokio::test]
-async fn test_list_locations_filter_construction_by_type() {
-    use unet_core::prelude::{Filter, FilterOperation, FilterValue};
-
-    let location_type = "datacenter".to_string();
-
-    // Test filter construction similar to list_locations function
-    let filter = Filter {
-        field: "location_type".to_owned(),
-        operation: FilterOperation::Equals,
-        value: FilterValue::String(location_type.clone()),
-    };
-
-    assert_eq!(filter.field, "location_type");
-    assert!(matches!(filter.operation, FilterOperation::Equals));
-    match filter.value {
-        FilterValue::String(value) => assert_eq!(value, location_type),
-        _ => panic!("Expected String filter value"),
-    }
-}
-
-#[tokio::test]
-async fn test_list_locations_filter_construction_by_parent() {
-    use unet_core::prelude::{Filter, FilterOperation, FilterValue};
-
-    let parent_id = Uuid::new_v4();
-
-    // Test filter construction similar to list_locations function
-    let filter = Filter {
-        field: "parent_id".to_owned(),
-        operation: FilterOperation::Equals,
-        value: FilterValue::Uuid(parent_id),
-    };
-
-    assert_eq!(filter.field, "parent_id");
-    assert!(matches!(filter.operation, FilterOperation::Equals));
-    match filter.value {
-        FilterValue::Uuid(id) => assert_eq!(id, parent_id),
-        _ => panic!("Expected UUID filter value"),
-    }
-}
-
-#[tokio::test]
-async fn test_list_locations_sort_construction() {
-    use unet_core::prelude::{Sort, SortDirection};
-
-    // Test sort construction similar to list_locations function
-    let sort = Sort {
-        field: "name".to_owned(),
-        direction: SortDirection::Ascending,
-    };
-
-    assert_eq!(sort.field, "name");
-    assert!(matches!(sort.direction, SortDirection::Ascending));
-}
-
-#[tokio::test]
-async fn test_list_locations_query_options_construction() {
-    use unet_core::prelude::{
-        Filter, FilterOperation, FilterValue, Pagination, QueryOptions, Sort, SortDirection,
-    };
-
-    let location_type = "datacenter".to_string();
-    let parent_id = Uuid::new_v4();
-    let page = 2_u64;
-    let per_page = 30_u64;
-
-    // Construct QueryOptions similar to list_locations function
-    let filters = vec![
-        Filter {
-            field: "location_type".to_owned(),
-            operation: FilterOperation::Equals,
-            value: FilterValue::String(location_type),
-        },
-        Filter {
-            field: "parent_id".to_owned(),
-            operation: FilterOperation::Equals,
-            value: FilterValue::Uuid(parent_id),
-        },
-    ];
-
-    let sort = vec![Sort {
-        field: "name".to_owned(),
-        direction: SortDirection::Ascending,
-    }];
-
-    let offset = usize::try_from((page - 1) * per_page).unwrap();
-    let limit = usize::try_from(per_page).unwrap();
-
-    let pagination = Some(Pagination { offset, limit });
-
-    let options = QueryOptions {
-        filters,
-        sort,
-        pagination,
-    };
-
-    // Verify construction
-    assert_eq!(options.filters.len(), 2);
-    assert_eq!(options.sort.len(), 1);
-    assert!(options.pagination.is_some());
-
-    if let Some(ref pagination) = options.pagination {
-        assert_eq!(pagination.offset, 30); // (2-1) * 30 = 30
-        assert_eq!(pagination.limit, 30);
-    }
-}
-
-// UPDATE LOCATION ARGUMENT VALIDATION TESTS
-
-#[tokio::test]
-async fn test_update_location_partial_updates() {
-    let location_id = Uuid::new_v4();
-
-    // Test that individual fields can be updated
-    let args = UpdateLocationArgs {
-        id: location_id,
-        name: Some("updated-name".to_string()),
-        location_type: None,
-        parent_id: None,
-        address: None,
-        city: None,
-        country: None,
-        custom_data: None,
-    };
-
-    // Verify only name field is set for update
-    assert_eq!(args.id, location_id);
-    assert_eq!(args.name, Some("updated-name".to_string()));
-    assert_eq!(args.location_type, None);
-    assert_eq!(args.parent_id, None);
-    assert_eq!(args.address, None);
-    assert_eq!(args.city, None);
-    assert_eq!(args.country, None);
-    assert_eq!(args.custom_data, None);
-}
-
-#[tokio::test]
-async fn test_update_location_address_combination() {
-    let address = Some("456 New Ave".to_string());
-    let city = Some("Boston".to_string());
-    let country = Some("USA".to_string());
-
-    // Test address combination logic from update_location
-    let mut address_parts = Vec::new();
-    if let Some(addr) = address {
-        address_parts.push(addr);
-    }
-    if let Some(c) = city {
-        address_parts.push(c);
-    }
-    if let Some(co) = country {
-        address_parts.push(co);
-    }
-
-    let combined_address = if address_parts.is_empty() {
-        None
-    } else {
-        Some(address_parts.join(", "))
-    };
-
-    assert_eq!(
-        combined_address,
-        Some("456 New Ave, Boston, USA".to_string())
-    );
-}
-
-#[tokio::test]
-async fn test_update_location_custom_data_parsing() {
-    let custom_data_str = r#"{"environment": "staging", "capacity": 75}"#;
-    let result = serde_json::from_str::<JsonValue>(custom_data_str);
-
-    assert!(result.is_ok());
-    let value = result.unwrap();
-    assert_eq!(value["environment"], "staging");
-    assert_eq!(value["capacity"], 75);
-}
-
-// DELETE LOCATION CONFIRMATION TESTS
-
-#[tokio::test]
-async fn test_delete_location_confirmation_logic() {
-    // Test confirmation logic patterns
-    let input_variations = vec![
-        ("y", true),
-        ("Y", true),
-        ("yes", true),
-        ("YES", true),
-        ("Yes", true),
-        ("n", false),
-        ("N", false),
-        ("no", false),
-        ("NO", false),
-        ("No", false),
-        ("", false),
-        ("maybe", false),
-        ("quit", false),
-    ];
-
-    for (input, expected) in input_variations {
-        let should_proceed = input.trim().to_lowercase().starts_with('y');
-        assert_eq!(
-            should_proceed, expected,
-            "Input '{input}' should return {expected}"
-        );
-    }
-}
-
-#[tokio::test]
-async fn test_delete_location_yes_flag_bypass() {
-    let location_id = Uuid::new_v4();
-
-    let args_with_yes = DeleteLocationArgs {
-        id: location_id,
-        yes: true,
-    };
-
-    let args_without_yes = DeleteLocationArgs {
-        id: location_id,
-        yes: false,
-    };
-
-    // When yes=true, no confirmation should be needed
-    assert!(args_with_yes.yes);
-    // When yes=false, confirmation should be required
-    assert!(!args_without_yes.yes);
-}
+#[path = "crud_business_logic_tests/add_address_tests.rs"]
+mod add_address_tests;
+#[path = "crud_business_logic_tests/json_builder_tests.rs"]
+mod json_builder_tests;
+#[path = "crud_business_logic_tests/query_update_delete_tests.rs"]
+mod query_update_delete_tests;
